+++ main.go
@@ -97,9 +97,6 @@ func (c *LargeObjectsPlugin) Run(cliConnection plugin.CliConnection, args []stri
 
 	// Associate each subcommand with a handler function
 	c.subcommands = map[string](command){
-		// Help command
-		//helpCommand: c.help,
-
 		// Authenticate command
 		getAuthInfoCommand: command{
 			name:            getAuthInfoCommand,
@@ -210,9 +207,8 @@ func (c *LargeObjectsPlugin) Run(cliConnection plugin.CliConnection, args []stri
 
 	// Dispatch the subcommand that the user wanted, if it exists
 	var err error
-	if len(args) < 2 {
-		err = fmt.Errorf("Please provide a valid subcommand\n" +
-			"A list of subcommands can be found with the command 'cf help os'")
+	if len(args) < 2 || args[1] == helpCommand {
+		err = c.help(args)
 	} else {
 		subcommand, found := c.subcommands[args[1]]
 		if !found {
@@ -258,251 +254,6 @@ func (c *LargeObjectsPlugin) executeCommand(cmd command, args []string) error {
 	fmt.Print(result)
 
 	return nil
-}
-
-// help prints help info for a given subcommand
-func (c *LargeObjectsPlugin) help(cliConnection plugin.CliConnection, args []string) error {
-	if len(args) < 2 {
-		help := "Please provide a subcommand to fetch its help info\n" +
-			"Available subcommands:\n" +
-			"      " + getAuthInfoCommand + "\n" +
-			"      " + showContainersCommand + "\n" +
-			"      " + containerInfoCommand + "\n" +
-			"      " + makeContainerCommand + "\n" +
-			"      " + updateContainerCommand + "\n" +
-			"      " + renameContainerCommand + "\n" +
-			"      " + deleteContainerCommand + "\n" +
-			"      " + showObjectsCommand + "\n" +
-			"      " + objectInfoCommand + "\n" +
-			"      " + putObjectCommand + "\n" +
-			"      " + getObjectCommand + "\n" +
-			"      " + renameObjectCommand + "\n" +
-			"      " + copyObjectCommand + "\n" +
-			"      " + deleteObjectCommand + "\n" +
-			"      " + makeDLOCommand + "\n" +
-			"      " + makeSLOCommand + "\n"
-
-		fmt.Print(help)
-
-		return nil
-	}
-
-	help, err := getSubcommandHelp(args[1])
-	if err != nil {
-		return err
-	}
-	fmt.Print(help)
-
-	return nil
-}
-
-func getSubcommandHelp(name string) (string, error) {
-	var subcommandMap = map[string]int{
-		getAuthInfoCommand:     0,
-		showContainersCommand:  1,
-		containerInfoCommand:   2,
-		makeContainerCommand:   3,
-		updateContainerCommand: 4,
-		renameContainerCommand: 5,
-		deleteContainerCommand: 6,
-		showObjectsCommand:     7,
-		objectInfoCommand:      8,
-		putObjectCommand:       9,
-		getObjectCommand:       10,
-		renameObjectCommand:    11,
-		copyObjectCommand:      12,
-		deleteObjectCommand:    13,
-		makeDLOCommand:         14,
-		makeSLOCommand:         15,
-	}
-
-	var subcommands = []plugin.Command{
-		{
-			Name:     getAuthInfoCommand,
-			HelpText: "Authenticate with Object Storage and save credentials",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + getAuthInfoCommand +
-					" service_name [--url] [-x]",
-				Options: map[string]string{
-					"url": "Display auth url in quiet mode",
-					"x":   "Display x-auth token in quiet mode",
-				},
-			},
-		},
-		{
-			Name:     showContainersCommand,
-			HelpText: "Show all containers in an Object Storage instance",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + showContainersCommand +
-					" service_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     containerInfoCommand,
-			HelpText: "Show a given container's information",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + containerInfoCommand +
-					" service_name container_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     makeContainerCommand,
-			HelpText: "Create a new container in an Object Storage instance",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + makeContainerCommand +
-					" service_name container_name [headers...] [r] [-r]",
-				Options: map[string]string{
-					"r":  "Short name for global read header",
-					"-r": "Short name for remove read restrictions header",
-				},
-			},
-		},
-		{
-			Name:     updateContainerCommand,
-			HelpText: "Update a container's metadata",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + updateContainerCommand +
-					" service_name container_name headers... [r] [-r]",
-				Options: map[string]string{
-					"r":  "Short name for global read header",
-					"-r": "Short name for remove read restrictions header",
-				},
-			},
-		},
-		{
-			Name:     renameContainerCommand,
-			HelpText: "Rename a container",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + renameContainerCommand +
-					" service_name container_name new_container_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     deleteContainerCommand,
-			HelpText: "Remove a container from an Object Storage instance",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + deleteContainerCommand +
-					" service_name container_name [-f]",
-				Options: map[string]string{
-					"f": "Force delete even if not empty",
-				},
-			},
-		},
-		{
-			Name:     showObjectsCommand,
-			HelpText: "Show all objects in a container",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + showObjectsCommand +
-					" service_name container_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     objectInfoCommand,
-			HelpText: "Show a given object's information",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + objectInfoCommand +
-					" service_name container_name object_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     putObjectCommand,
-			HelpText: "Upload a file as an object to Object Storage",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + putObjectCommand +
-					" service_name container_name path_to_local_file [-n object_name]",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     getObjectCommand,
-			HelpText: "Download an object from Object Storage",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + getObjectCommand +
-					" service_name container_name object_name path_to_dl_location",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     renameObjectCommand,
-			HelpText: "Rename an object",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + renameObjectCommand +
-					" service_name container_name object_name new_object_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     copyObjectCommand,
-			HelpText: "Copy an object to another container",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + copyObjectCommand +
-					" service_name container_name object_name new_container_name",
-				Options: map[string]string{},
-			},
-		},
-		{
-			Name:     deleteObjectCommand,
-			HelpText: "Remove an object from a container",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + deleteObjectCommand +
-					" service_name container_name object_name [-l]",
-				Options: map[string]string{
-					"l": "Delete all files associated with large object manifest object_name",
-				},
-			},
-		},
-		{
-			Name:     makeDLOCommand,
-			HelpText: "Create a Dynamic Large Object in Object Storage",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + makeDLOCommand +
-					" service_name dlo_container dlo_name [-c object_container] [-p dlo_prefix]",
-				Options: map[string]string{
-					"c": "Destination container for DLO segments (defaults to dlo_container)",
-					"p": "Prefix to be used for DLO segments (default to dlo_name)",
-				},
-			},
-		},
-		{
-			Name:     makeSLOCommand,
-			HelpText: "Create a Static Large Object in Object Storage",
-			UsageDetails: plugin.Usage{
-				Usage: "cf " + namespace + " " + makeSLOCommand +
-					" service_name slo_container slo_name source_file [-m] [-o output_file] [-s chunk_size] [-j num_threads]",
-				Options: map[string]string{
-					"m": "Only upload missing chunks",
-					"o": "Destination for log data, if desired",
-					"s": "Chunk size, in bytes (defaults to create 1GB chunks)",
-					"j": "Maximum number of uploader threads (defaults to the available number of CPUs)",
-				},
-			},
-		},
-	}
-
-	idx, found := subcommandMap[name]
-	if !found {
-		return "", fmt.Errorf("%s is not a valid subcommand", name)
-	}
-
-	help := ""
-	help += fmt.Sprintf("NAME:\n\t%s - %s\n\nUSAGE:\n\t%s\n", subcommands[idx].Name, subcommands[idx].HelpText, subcommands[idx].UsageDetails.Usage)
-	if len(subcommands[idx].UsageDetails.Options) > 0 {
-		help += fmt.Sprintf("\nOPTIONS:\n")
-		for k, v := range subcommands[idx].UsageDetails.Options {
-			if name != makeContainerCommand && name != updateContainerCommand {
-				help += fmt.Sprintf("\t-%s\t\t%v\n", k, v)
-			} else {
-				help += fmt.Sprintf("\t%s\t\t%v\n", k, v)
-			}
-		}
-	}
-
-	return help, nil
 }
 
 // GetMetadata returns a PluginMetadata struct with information
